@@ -1,124 +1,61 @@
 name: ecommerce_app
-description: "E-Commerce App with Firebase"
-# The following line prevents the package from being accidentally published to
-# pub.dev using `flutter pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
-
+description: "A new Flutter project."
+publish_to: 'none'
 
 version: 1.0.0+1
 
-
 environment:
-  sdk: '>=3.0.0 <4.0.0'
-
+  sdk: ^3.8.1
 
 dependencies:
   flutter:
     sdk: flutter
 
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
-
-
-  # Firebase packages
-  firebase_core: ^4.2.0
-  firebase_auth: ^6.1.1
-  cloud_firestore: ^6.0.3
+  firebase_core: ^2.24.0
+  flutter_native_splash: 2.4.0
+  firebase_auth: ^4.17.3
+  cloud_firestore: ^4.13.0
   provider: ^6.1.5+1
+  google_fonts: ^6.3.2
   intl: ^0.20.2
-
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
+  flutter_lints: ^6.0.0
+  flutter_launcher_icons: ^0.13.1 # I have added this package
 
-  # The "flutter_lints" package below contains a set of recommended lints to
-  # encourage good coding practices. The lint set provided by the package is
-  # activated in the `analysis_options.yaml` file located at the root of your
-  # package. See that file for information about deactivating specific lint
-  # rules and activating additional ones.
-  flutter_lints: ^4.0.0
-
-
-  # Flutter native splash
-  flutter_native_splash: ^2.4.0
-
-
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-
-# The following section is specific to Flutter packages.
 flutter:
-
-
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
   uses-material-design: true
-
-
-  # Add assets
   assets:
-    - assets/images/
+    - assets/images/splash_logo.png
+    - assets/images/app_logo.png
+
+# Configuration for flutter_launcher_icons
+flutter_launcher_icons:
+  android: "launcher_icon"
+  ios: true
+  image_path: "assets/images/app_logo.png"
+  min_sdk_android: 21 # optional
 
 
-  # To add assets to your application, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
-
-
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/to/resolution-aware-images
-
-
-  # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev/to/asset-from-package
-
-
-  # To add custom fonts to your application, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/to/font-from-package
-
-
-# 1. Add this new section at the end of the file
 flutter_native_splash:
-
 
   # 2. Set the background color of your splash screen
   color: "#FFFFFF" # This is white
 
-
   # 3. Point to your splash logo file
-  image: assets/images/tulip_logo.png
-
+  image: assets/images/splash_logo.png
 
   # 4. Configuration for the Android splash screen
   android_12:
     color: "#FFFFFF"
-    image: assets/images/tulip_logo.png
-
+    image: assets/images/splash_logo.png
 
   # 5. Configuration for the web splash screen
+
   # Web: only supports `true` (boolean), no image/color override
+
   web: true
