@@ -14,12 +14,27 @@ class _AdminOrderScreenState extends State<AdminOrderScreen> {
   final FirebaseFirestore _firestore = FirebaseFirestore.instance;
 
   // 2. This is the function that updates the status in Firestore
-  Future<void> _updateOrderStatus(String orderId, String newStatus) async {
+  // 1. MODIFY this function to accept userId
+  Future<void> _updateOrderStatus(String orderId, String newStatus, String userId) async {
     try {
-      // 3. Find the document and update the 'status' field
+      // 3. This part is the same (update the order)
       await _firestore.collection('orders').doc(orderId).update({
         'status': newStatus,
       });
+
+      // 3. --- ADD THIS NEW LOGIC ---
+      //    Create a new notification document
+      await _firestore.collection('notifications').add({
+        'userId': userId, // 4. The user this notification is for
+        'title': 'Order Status Updated',
+        'body': 'Your order ($orderId) has been updated to "$newStatus".',
+        'orderId': orderId,
+        'createdAt': FieldValue.serverTimestamp(),
+        'isRead': false, // 5. Mark it as unread
+      });
+      // --- END OF NEW LOGIC ---
+
+      if (!mounted) return;
       ScaffoldMessenger.of(context).showSnackBar(
         const SnackBar(content: Text('Order status updated!')),
       );
@@ -31,10 +46,11 @@ class _AdminOrderScreenState extends State<AdminOrderScreen> {
   }
 
   // 4. This function shows the update dialog
-  void _showStatusDialog(String orderId, String currentStatus) {
+  // 1. MODIFY this function to accept userId
+  void _showStatusDialog(String orderId, String currentStatus, String userId) {
     showDialog(
       context: context,
-      builder: (context) {
+      builder: (dialogContext) {
         // 5. A list of all possible statuses
         const statuses = ['Pending', 'Processing', 'Shipped', 'Delivered', 'Cancelled'];
 
@@ -50,15 +66,16 @@ class _AdminOrderScreenState extends State<AdminOrderScreen> {
                 trailing: currentStatus == status ? const Icon(Icons.check) : null,
                 onTap: () {
                   // 8. When tapped:
-                  _updateOrderStatus(orderId, status); // Call update
-                  Navigator.of(context).pop(); // Close the dialog
+                  // 2. PASS userId to our update function
+                  _updateOrderStatus(orderId, status, userId); // Call update
+                  Navigator.of(dialogContext).pop(); // Close the dialog
                 },
               );
             }).toList(),
           ),
           actions: [
             TextButton(
-              onPressed: () => Navigator.of(context).pop(),
+              onPressed: () => Navigator.of(dialogContext).pop(),
               child: const Text('Close'),
             )
           ],
@@ -110,6 +127,9 @@ class _AdminOrderScreenState extends State<AdminOrderScreen> {
               // 6. Get the current status
               final String status = orderData['status'];
 
+              // 3. PASS userId from the order data to our dialog
+              final String userId = orderData['userId'] ?? 'Unknown User';
+
               // 7. Build a Card for each order
               return Card(
                 margin: const EdgeInsets.all(8.0),
@@ -119,7 +139,7 @@ class _AdminOrderScreenState extends State<AdminOrderScreen> {
                     style: const TextStyle(fontWeight: FontWeight.bold, fontSize: 12),
                   ),
                   subtitle: Text(
-                      'User: ${orderData['userId']}\n'
+                      'User: $userId\n'
                           'Total: ₱${(orderData['totalPrice']).toStringAsFixed(2)} | Date: $formattedDate'
                   ),
                   isThreeLine: true,
@@ -139,7 +159,8 @@ class _AdminOrderScreenState extends State<AdminOrderScreen> {
 
                   // 9. On tap, show our update dialog
                   onTap: () {
-                    _showStatusDialog(order.id, status);
+                    // 3. PASS userId from the order data to our dialog
+                    _showStatusDialog(order.id, status, userId);
                   },
                 ),
               );
