@@ -1,25 +1,26 @@
-// Part 1: Imports
+import 'package:ecommerce_app/Widgets/notification_icon.dart';
+import 'package:ecommerce_app/providers/cart_provider.dart';
+import 'package:ecommerce_app/screens/admin_panel_screen.dart';
+import 'package:ecommerce_app/screens/cart_screen.dart';
+import 'package:ecommerce_app/screens/chat_screen.dart';
+import 'package:ecommerce_app/screens/order_history_screen.dart';
+import 'package:ecommerce_app/screens/profile_screen.dart';
+import 'package:ecommerce_app/widgets/product_card.dart';
+import 'package:ecommerce_app/screens/product_detail_screen.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:ecommerce_app/screens/admin_panel_screen.dart';
-import 'package:ecommerce_app/widgets/product_card.dart';
-import 'package:ecommerce_app/screens/product_detail_screen.dart'; // 1. ADD THIS IMPORT
-import 'package:ecommerce_app/providers/cart_provider.dart'; // 1. ADD THIS
-import 'package:ecommerce_app/screens/cart_screen.dart'; // 2. ADD THIS
-import 'package:provider/provider.dart'; // 3. ADD THIS
-import 'package:ecommerce_app/screens/order_history_screen.dart'; // 1. ADD THIS
+import 'package:provider/provider.dart';
 
 class HomeScreen extends StatefulWidget {
   const HomeScreen({super.key});
-
-  @override
   State<HomeScreen> createState() => _HomeScreenState();
 }
 
 class _HomeScreenState extends State<HomeScreen> {
-  String _userRole = 'admin';
+  String _userRole = 'user';
   final User? _currentUser = FirebaseAuth.instance.currentUser;
+  final FirebaseFirestore _firestore = FirebaseFirestore.instance;
 
   @override
   void initState() {
@@ -34,7 +35,6 @@ class _HomeScreenState extends State<HomeScreen> {
           .collection('users')
           .doc(_currentUser!.uid)
           .get();
-
       if (doc.exists && doc.data() != null) {
         setState(() {
           _userRole = doc.data()!['role'];
@@ -57,25 +57,19 @@ class _HomeScreenState extends State<HomeScreen> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text(_currentUser != null ? 'Welcome, ${_currentUser!.email}' : 'Home'),
+        title: Image.asset(
+          'assets/images/app_logo.png',
+          height: 40,
+        ),
         actions: [
-
-          // 1. --- ADD THIS NEW WIDGET ---
-          // This is a special, efficient way to use Provider
           Consumer<CartProvider>(
-            // 2. The "builder" function rebuilds *only* the icon
             builder: (context, cart, child) {
-              // 3. The "Badge" widget adds a small label
               return Badge(
-                // 4. Get the count from the provider
                 label: Text(cart.itemCount.toString()),
-                // 5. Only show the badge if the count is > 0
                 isLabelVisible: cart.itemCount > 0,
-                // 6. This is the child (our icon button)
                 child: IconButton(
                   icon: const Icon(Icons.shopping_cart),
                   onPressed: () {
-                    // 7. Navigate to the CartScreen
                     Navigator.of(context).push(
                       MaterialPageRoute(
                         builder: (context) => const CartScreen(),
@@ -86,46 +80,42 @@ class _HomeScreenState extends State<HomeScreen> {
               );
             },
           ),
-
-          // 2. --- ADD THIS NEW BUTTON ---
+          const NotificationIcon(),
           IconButton(
-            icon: const Icon(Icons.receipt_long), // A "receipt" icon
-            tooltip: 'My Orders',
-            onPressed: () {
-              Navigator.of(context).push(
-                MaterialPageRoute(
-                  builder: (context) => const OrderHistoryScreen(),
-                ),
-              );
-            },
-          ),
-
-          if (_userRole == 'admin')
-            IconButton(
-              icon: const Icon(Icons.admin_panel_settings),
-              tooltip: 'Admin Panel',
+              icon: const Icon(Icons.receipt_long),
               onPressed: () {
                 Navigator.of(context).push(
                   MaterialPageRoute(
-                    builder: (context) => const AdminPanelScreen(),
+                    builder: (context) => const OrderHistoryScreen(),
                   ),
                 );
-              },
-            ),
+              }),
+          if (_userRole == 'admin')
+            IconButton(
+                icon: const Icon(Icons.admin_panel_settings),
+                onPressed: () {
+                  Navigator.of(context).push(
+                    MaterialPageRoute(
+                      builder: (context) => const AdminPanelScreen(),
+                    ),
+                  );
+                }),
           IconButton(
-            icon: const Icon(Icons.logout),
-            tooltip: 'Logout',
-            onPressed: _signOut,
-          ),
+              icon: const Icon(Icons.person_outline),
+              onPressed: () {
+                Navigator.of(context).push(
+                  MaterialPageRoute(
+                    builder: (context) => const ProfileScreen(),
+                  ),
+                );
+              }),
         ],
       ),
-
       body: StreamBuilder<QuerySnapshot>(
         stream: FirebaseFirestore.instance
             .collection('products')
             .orderBy('createdAt', descending: true)
             .snapshots(),
-
         builder: (context, snapshot) {
           if (snapshot.connectionState == ConnectionState.waiting) {
             return const Center(child: CircularProgressIndicator());
@@ -151,30 +141,21 @@ class _HomeScreenState extends State<HomeScreen> {
               mainAxisSpacing: 10,
               childAspectRatio: 3 / 4,
             ),
-
             itemCount: products.length,
             itemBuilder: (context, index) {
-              // 1. Get the whole document
               final productDoc = products[index];
-              // 2. Get the data map
               final productData = productDoc.data() as Map<String, dynamic>;
 
-              // 3. Find your old ProductCard
               return ProductCard(
                 productName: productData['name'],
                 price: productData['price'],
                 imageUrl: productData['imageUrl'],
-
-                // 4. --- THIS IS THE NEW PART ---
-                //    Add the onTap property
                 onTap: () {
-                  // 5. Navigate to the new screen
                   Navigator.of(context).push(
                     MaterialPageRoute(
                       builder: (context) => ProductDetailScreen(
-                        // 6. Pass the data to the new screen
                         productData: productData,
-                        productId: productDoc.id, // 7. Pass the unique ID!
+                        productId: productDoc.id,
                       ),
                     ),
                   );
@@ -184,6 +165,42 @@ class _HomeScreenState extends State<HomeScreen> {
           );
         },
       ),
+      floatingActionButton: _userRole == 'user'
+          ? StreamBuilder<DocumentSnapshot>(
+        stream: _firestore
+            .collection('chats')
+            .doc(_currentUser!.uid)
+            .snapshots(),
+        builder: (context, snapshot) {
+          int unreadCount = 0;
+          if (snapshot.hasData && snapshot.data!.exists) {
+            final data = snapshot.data!.data();
+            if (data != null) {
+              unreadCount =
+                  (data as Map<String, dynamic>)['unreadByUserCount'] ?? 0;
+            }
+          }
+
+          return Badge(
+            label: Text('$unreadCount'),
+            isLabelVisible: unreadCount > 0,
+            child: FloatingActionButton.extended(
+              icon: const Icon(Icons.support_agent),
+              label: const Text('Contact Admin'),
+              onPressed: () {
+                Navigator.of(context).push(
+                  MaterialPageRoute(
+                    builder: (context) => ChatScreen(
+                      chatRoomId: _currentUser!.uid,
+                    ),
+                  ),
+                );
+              },
+            ),
+          );
+        },
+      )
+          : null,
     );
   }
 }
