@@ -12,24 +12,19 @@ class OrderHistoryScreen extends StatelessWidget {
     final User? user = FirebaseAuth.instance.currentUser;
 
     return Scaffold(
-      appBar: AppBar(
-        title: const Text('My Orders'),
-      ),
+      appBar: AppBar(title: const Text('My Orders')),
       // 3. Check if the user is logged in
       body: user == null
-          ? const Center(
-        child: Text('Please log in to see your orders.'),
-      )
+          ? const Center(child: Text('Please log in to see your orders.'))
       // 4. If logged in, show the StreamBuilder
           : StreamBuilder<QuerySnapshot>(
-
-        // 5. --- TEMPORARY FIX: REMOVE ORDERBY WHILE INDEX IS BEING CREATED ---
+        // 5. --- THIS IS THE CRITICAL NEW QUERY ---
         stream: FirebaseFirestore.instance
             .collection('orders')
         // 6. Filter the 'orders' collection
             .where('userId', isEqualTo: user.uid)
-        // 7. TEMPORARILY REMOVED: Sort by date, newest first
-        // .orderBy('createdAt', descending: true) // COMMENT OUT THIS LINE FOR NOW
+            .orderBy('createdAt', descending: true)
+        // 7. Sort by date, newest first
             .snapshots(),
 
         builder: (context, snapshot) {
@@ -51,26 +46,18 @@ class OrderHistoryScreen extends StatelessWidget {
           }
 
           // 11. We have data! Get the list of order documents
+          // DEBUG: Print what we're getting
           final orderDocs = snapshot.data!.docs;
 
-          // 12. --- ADD LOCAL SORTING INSTEAD OF FIRESTORE SORTING ---
-          // Sort the documents locally by createdAt in descending order
-          orderDocs.sort((a, b) {
-            final aData = a.data() as Map<String, dynamic>;
-            final bData = b.data() as Map<String, dynamic>;
-            final aTime = (aData['createdAt'] as Timestamp).millisecondsSinceEpoch;
-            final bTime = (bData['createdAt'] as Timestamp).millisecondsSinceEpoch;
-            return bTime.compareTo(aTime); // Descending order (newest first)
-          });
-
-          // 13. Use ListView.builder to show the list
+          // 12. Use ListView.builder to show the list
           return ListView.builder(
             itemCount: orderDocs.length,
             itemBuilder: (context, index) {
-              // 14. Get the data for a single order
-              final orderData = orderDocs[index].data() as Map<String, dynamic>;
+              // 13. Get the data for a single order
+              final orderData =
+              orderDocs[index].data() as Map<String, dynamic>;
 
-              // 15. Return our custom OrderCard widget
+              // 14. Return our custom OrderCard widget
               return OrderCard(orderData: orderData);
             },
           );
