@@ -21,7 +21,7 @@ class OrderSuccessScreen extends StatelessWidget {
               // 2. A nice success icon
               const Icon(
                 Icons.check_circle_outline,
-                color: Colors.pinkAccent,
+                color: Colors.green,
                 size: 100,
               ),
               const SizedBox(height: 20),
