@@ -3,14 +3,11 @@ import 'package:ecommerce_app/providers/cart_provider.dart'; // 1. ADD THIS
 import 'package:provider/provider.dart'; // 2. ADD THIS
 
 // 1. This is a new StatelessWidget
-class ProductDetailScreen extends StatelessWidget {
-
-  // 2. We will pass in the product's data (the map)
+// 1. Change StatelessWidget to StatefulWidget
+class ProductDetailScreen extends StatefulWidget {
   final Map<String, dynamic> productData;
-  // 3. We'll also pass the unique product ID (critical for 'Add to Cart' later)
   final String productId;
 
-  // 4. The constructor takes both parameters
   const ProductDetailScreen({
     super.key,
     required this.productData,
@@ -18,31 +15,49 @@ class ProductDetailScreen extends StatelessWidget {
   });
 
   @override
+  // 2. Create the State class
+  State<ProductDetailScreen> createState() => _ProductDetailScreenState();
+}
+
+// 3. Rename the main class to _ProductDetailScreenState and extend State
+class _ProductDetailScreenState extends State<ProductDetailScreen> {
+  // 4. ADD OUR NEW STATE VARIABLE FOR QUANTITY
+  int _quantity = 1;
+
+  // 1. ADD THIS FUNCTION
+  void _incrementQuantity() {
+    setState(() {
+      _quantity++;
+    });
+  }
+
+  // 2. ADD THIS FUNCTION
+  void _decrementQuantity() {
+    // We don't want to go below 1
+    if (_quantity > 1) {
+      setState(() {
+        _quantity--;
+      });
+    }
+  }
+
+  // 5. The build method will go inside here
+  @override
   Widget build(BuildContext context) {
+    // 1. We now access productData using 'widget.'
+    final String name = widget.productData['name'];
+    final String description = widget.productData['description'];
+    final String imageUrl = widget.productData['imageUrl'];
+    final double price = widget.productData['price'];
 
-    // 1. Extract data from the map for easier use
-    final String name = productData['name'];
-    final String description = productData['description'];
-    final String imageUrl = productData['imageUrl'];
-    final double price = productData['price'];
-
-    // 1. ADD THIS LINE: Get the CartProvider
-    // We set listen: false because we are not rebuilding, just calling a function
+    // 2. Get the CartProvider (same as before)
     final cart = Provider.of<CartProvider>(context, listen: false);
 
-    // 2. The main screen widget
     return Scaffold(
-      appBar: AppBar(
-        // 3. Show the product name in the top bar
-        title: Text(name),
-      ),
-      // 4. This allows scrolling if the description is very long
+      // ... (AppBar is the same)
       body: SingleChildScrollView(
         child: Column(
-          // 5. Make children fill the width
-          crossAxisAlignment: CrossAxisAlignment.stretch,
           children: [
-
             // 6. The large product image
             Image.network(
               imageUrl,
@@ -65,12 +80,12 @@ class ProductDetailScreen extends StatelessWidget {
             ),
 
             // 8. A Padding widget to contain all the text
+            // 3. Find the "Padding" widget that holds your "Add to Cart" button
             Padding(
               padding: const EdgeInsets.all(16.0),
               child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
+                crossAxisAlignment: CrossAxisAlignment.stretch,
                 children: [
-
                   // 9. Product Name (large font)
                   Text(
                     name,
@@ -87,7 +102,7 @@ class ProductDetailScreen extends StatelessWidget {
                     style: TextStyle(
                       fontSize: 24,
                       fontWeight: FontWeight.w600,
-                      color: Colors.pink,
+                      color: Colors.deepPurple,
                     ),
                   ),
                   const SizedBox(height: 16),
@@ -111,19 +126,58 @@ class ProductDetailScreen extends StatelessWidget {
                   ),
                   const SizedBox(height: 30),
 
+                  // 4. --- ADD THIS NEW SECTION ---
+                  //    (before the "Add to Cart" button)
+                  const SizedBox(height: 20),
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: [
+                      // 5. DECREMENT BUTTON
+                      IconButton.filledTonal(
+                        icon: const Icon(Icons.remove),
+                        onPressed: _decrementQuantity,
+                      ),
+
+                      // 6. QUANTITY DISPLAY
+                      Padding(
+                        padding: const EdgeInsets.symmetric(horizontal: 20),
+                        child: Text(
+                          '$_quantity', // 7. Display our state variable
+                          style: const TextStyle(
+                            fontSize: 24,
+                            fontWeight: FontWeight.bold,
+                          ),
+                        ),
+                      ),
+
+                      // 8. INCREMENT BUTTON
+                      IconButton.filled(
+                        icon: const Icon(Icons.add),
+                        onPressed: _incrementQuantity,
+                      ),
+                    ],
+                  ),
+                  const SizedBox(height: 20),
+                  // --- END OF NEW SECTION ---
+
                   // 13. The "Add to Cart" button (UI ONLY)
                   // It doesn't do anything... yet.
                   ElevatedButton.icon(
                     onPressed: () {
-                      // 4. THIS IS THE NEW LOGIC!
-                      // Call the addItem function from our provider
-                      cart.addItem(productId, name, price);
+                      // 10. --- THIS IS THE UPDATED LOGIC ---
+                      //    We now pass the _quantity from our state
+                      cart.addItem(
+                        widget.productId,
+                        name,
+                        price,
+                        _quantity, // 11. Pass the selected quantity
+                      );
 
-                      // 5. Show a confirmation pop-up
+                      // 12. Update the SnackBar message
                       ScaffoldMessenger.of(context).showSnackBar(
-                        const SnackBar(
-                          content: Text('Added to cart!'),
-                          duration: Duration(seconds: 2),
+                        SnackBar(
+                          content: Text('Added $_quantity x $name to cart!'),
+                          duration: const Duration(seconds: 2),
                         ),
                       );
                     },
